+++ CloudNotes/CloudNotes/View/MemoListCell.swift
@@ -9,7 +9,6 @@ import UIKit
 
 class MemoListCell: UITableViewCell {
     static let identifier = "MemoListTableViewCell"
-    private var dateConverter = DateConverter()
     
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: .default , reuseIdentifier: MemoListCell.identifier)
@@ -73,10 +72,10 @@ class MemoListCell: UITableViewCell {
         ])
     }
     
-    func configureCell(data: Memo) {
-        title.text = data.title
-        body.text = data.body
-        lastModified.text = dateConverter.convertDate(date: data.lastModified)
+    func configureCell(memoData: Memo, stringLastModified: String) {
+        title.text = memoData.title
+        body.text = memoData.body
+        lastModified.text = stringLastModified
     }
     
 }
